Now supports 64bit machines using the ivm
This is the result of a _very_ quick and dirty attempt to prove
that psyco can be ported to x86_64, I'll need to go back and
decide how to really handle the 32/64 bit int issue - I've used
different types all over the place.

I'm committing the generated files in c/ivm/prolog/ because I
had to hack some of them by hand - The generated code isn't
very good - no intermediate assignment elimination makes
the code look more complex that it is - will probably rewrite
this by hand.

All the tests I'm adding here work apart from django_template.py
which raises a `KeyError: None` which I don't understand yet.

I'm very supprised by how easy this has been so far but there
is a lot still to do.

--HG--
branch : py27

+++ c/Objects/pobject.h
@@ -11,9 +11,9 @@
 
 /*#define OB_REFCOUNT         never directly manipulated*/
 #define OB_type      DEF_FIELD(PyObject, PyTypeObject*, ob_type, NO_PREV_FIELD)
-#define FIX_size     NONNEG_FIELD(PyVarObject,     int, ob_size, OB_type)
+#define FIX_size     NONNEG_FIELD(PyVarObject,     Py_ssize_t, ob_size, OB_type)
 #define VAR_size     FMUT(FIX_size)
-#define FIX_signed_size DEF_FIELD(PyVarObject,     int, ob_size, OB_type)
+#define FIX_signed_size DEF_FIELD(PyVarObject,     Py_ssize_t, ob_size, OB_type)
 #define VAR_signed_size FMUT(FIX_signed_size)
 
 #define iOB_TYPE   FIELD_INDEX(OB_type)
